fix: fix packaging due to sub-directory
Moving the published package to a subdirectory caused a few issues in
the publishing process. First, including the README in the library
documentation is difficult due to path differences during normal builds
vs during packaging. Second, we have to specify which package to publish
because the package is not at the repository root.